iio: staging: adc: max77660: add wakeup support from ADC
During suspend, ADC is put in continuous conversion with proper lower/
higher threshold and when the channel ADC conversion value cross these
threshold, it generates interrupt which cause the system to be wakeup.

This configuration is platform configuration and only be enabled when
proper data is passed from platform.

bug 1301017

Change-Id: Ief7d994bf79beb34efa687866ce9ba210fddd03d
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/243861
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>