ARM: tegra: emc: Derating support
Add support for derating via keeping two separate sets of EMC tables.
One set is nominal, the other set is derated. Based on the temperature
reported by the DRAM the EMC thermal driver can specify which set of
tables the EMC driver should use when swapping frequencies.

This patch also adds support for a more graduated response to rising
temperature. The DRAM reports 3 levels of refresh and derating
requirements:

  0x4: Refresh x2
  0x5: Refresh x4
  0x6: Refresh x4 + derating

The particular combination of refresh modification and derating is now
picked based on the particular level of throttling necessary.

This feature is not imlpemented in older chips. The old approach is
maintained - just a 4x refresh timing for all temperature throttle
states.

Change-Id: I997a7245ec8cbe4ef6aa279f31ed58da6cd8239e
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: http://git-master/r/252624
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Yu-Huan Hsu <yhsu@nvidia.com>