ARM: tegra: ceres: tune haptic platform data
Problem:
There is continuous max77660 vibrator buzz during stress tests at
some random points in time.

Observations and fix:
It is found that matching the parameters pattern_signal_period
and feedback_duty_cycle with values used on pluto vibrator max77665
does not show the above problem. Hence, changed the duty cycle to
less than 100% value.

Possible Reason:
It could be the max77660 vibrator is not recommended to be used with
100% duty cycle and pattern signal period.

bug 1308274

Change-Id: I9ca42a3ba395f2d84a6f1d6acf07838aa7141d34
Signed-off-by: Ishwarya Balaji Gururajan <igururajan@nvidia.com>
Signed-off-by: Bitan Biswas <bbiswas@nvidia.com>
Reviewed-on: http://git-master/r/247635
Reviewed-by: Mrutyunjay Sawant <msawant@nvidia.com>
Tested-by: Mrutyunjay Sawant <msawant@nvidia.com>