ARM: tegra: config PWRGOOD_MASK in SOC_THERM
In T148, battery OC is signalled by max77660 though PG_OC. But this same
pin is also used to signal CPU_PWRGOOD.

The PWRGOOD_MASK describes the behavior of the OC alarm wrt the
CPU_PWRGOOD signal. If SW sets this bit to 1, it means that OC alarm
will not be asserted until CPU_PWRGOOD is also asserted.

This patch lets the throttle vector to specify the power-good mask.

Bug 1313669

Change-Id: If12d868de181ab6f0581686763ddcb748819301c
Signed-off-by: Sivaram Nair <sivaramn@nvidia.com>
Reviewed-on: http://git-master/r/250308
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Diwakar Tundlam <dtundlam@nvidia.com>