ARM: tegra: bbc: set percentage of bw for r/w
The BBC ISO bandwidth request includes BBCR and BBCW port bw.
A percentage of this bw should be designated for reads and writes
when calling the LA API so that the LA/PTSA settings can be updated
as needed. We are starting with a 50% split and can tune it in the
future.

Bug 1327858

Change-Id: I4acf4df5c130a1445ac3b5fa5e95dd8ff26e2df2
Signed-off-by: Neil Patel <neilp@nvidia.com>
Reviewed-on: http://git-master/r/250396
Reviewed-by: Mrutyunjay Sawant <msawant@nvidia.com>
Tested-by: Mrutyunjay Sawant <msawant@nvidia.com>