ARM: tegra14: clock: Update host1x bus operations
- Allowed low rates source pll selection for rates above low / high
threshold if it provides better approximation of target rate (kept
low rates pll as a sole source of rates below threshold).

- Added directional up-down rounding support

- Adjusted rounding algorithm to always round to closest rate within
+/-1 HZ range regardless of selected rounding direction (up or down).
This is done to avoid jumps across ladder of divider values when
parent rate is not an exact multiple of the divider.

- When switching from high rates source pll to low rates source pll,
skipped attempt to reduce intermittent divider, if it is already at
minimum possible for high rates source pll level (could not happen
anyway).

Change-Id: I41d39e6393216d849f2d58ae0cfdf15fb2be351a
Signed-off-by: Alex Frid <afrid@nvidia.com>
Reviewed-on: http://git-master/r/251165
Reviewed-by: Mrutyunjay Sawant <msawant@nvidia.com>
Tested-by: Mrutyunjay Sawant <msawant@nvidia.com>