ARM: tegra: mc: Make error handling make more sense
On T114 and later the errors being reported by the MC are starting
to grow in complexity. This takes that complexity and boils it down
into something more manageable for developers debugging SMMU and
other faults generated by the MC.

The aim of this patch is to make the errors more understandable and
correctly reflect the errors actually generated. The previous
approach to error handling was mis-representing certain errors.

Change-Id: I48a587aeed150a3ed998267f6feedd4afb2d35da
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: http://git-master/r/255282
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>