arm:tegra14: keep PLL-C ON if needed on voice call
The LP1BB entry logic that checks the EMC source to determine
if PLL-C must be ON in LP1BB must also apply to the case when
there is an ongoing voice call. Failing to do so results in
PLL-C being cut unconditionally, which may result in baseband
crashes.

bug 1331131

Change-Id: Iee416b3f1f1e030ae3f578f216ea372827271c94
Signed-off-by: Greg Heinrich <gheinrich@nvidia.com>
Reviewed-on: http://git-master/r/253280
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Herve Fache <hfache@nvidia.com>
Reviewed-by: Julien Vuillaumier <jvuillaumier@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Seshendra Gadagottu <sgadagottu@nvidia.com>
Reviewed-by: Aleksandr Frid <afrid@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>