ARM: tegra14: BGBIAS cells cannot be in DPD when entering LP0
For low frequencies that use Schmitt mode triggering in the EMC the
BGBIAS cells are not needed and can be turned off. However, if the
system is at a low frequency and goes into LP0 (i.e when the BBC is
in airplane mode) the bootrom will fail to read memory on wake up.
This occurs because the bootrom boots at the max EMC frequency (787
or 921 MHz) which requires the BGBIAS cells to be alive.

To fix the problem the BGBIAS cells are taken out of DPD when the
system goes into LP0.

Bug 1332120

Change-Id: I77e5a8fa48a2ea6dd6175fedc2280e368ef1126e
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: http://git-master/r/256523
Reviewed-by: Mrutyunjay Sawant <msawant@nvidia.com>
Tested-by: Mrutyunjay Sawant <msawant@nvidia.com>