arm: tegra: bypass PllP during LP1 suspend
RAM repair requires PllP, so it shouldn't be disabled. To save
power, instead of keeping it running at 408Mhz, enable bypass
mode, so RAM repair logic can be clocked by oscillator. This
is done when LP1 entry is from fast cluster only.

In addition, change PLLP_OUT0_RATIO to 0 so the reshift clock
is not being further divided down, change it back to default
value after PllP is enabled and bypass is disabled.

bug 1373419

Change-Id: I2bbcd15eaa0b222faadcae00448bd677c8387c69
Signed-off-by: Bo Yan <byan@nvidia.com>
Signed-off-by: Seshendra Gadagottu <sgadagottu@nvidia.com>
Reviewed-on: http://git-master/r/325703
Reviewed-by: Harshada Kale <hkale@nvidia.com>
Tested-by: Harshada Kale <hkale@nvidia.com>