ARM: tegra14: emc: Update EMC clock change
For low frequency transitions the AUTO_CAL process
requires the BGBIAS HW to be out of DPD. This change
makes sure the DPD settings for BGBIAS are cleared
for all clock changes and that auto calibration is
performed. Then for low frequencies the BGBIAS HW is
palced back in DPD and for high frequencies the auto
cal interval is restored. For low frequencies auto
cal is only done once.

Bug 1397645
Bug 1340437

Change-Id: I767e9ee5f646572b5479d1203c77588b767aa2cc
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: http://git-master/r/327880
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Aleksandr Frid <afrid@nvidia.com>
Reviewed-by: Yu-Huan Hsu <yhsu@nvidia.com>