video: tegra: nvmap: introduce iommu area flag
This patch adds the ability to keep the internal buffer(area) state
consistent. If invalid request comes, this returns an error with stack
dump. If client has some inconsistent usage of calls under heavy race
condition, this would catch and show that, letting kernel to be still
alive. The overhead of this is enough small to ignore from overall
perf.

Bug 1416207

Change-Id: Id0a5c3df69fd3df65b414a2ca23918f869a4a88c
Signed-off-by: Hiroshi Doyu <hdoyu@nvidia.com>
Reviewed-on: http://git-master/r/345574
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit