ARM: tegra: bb: use level triggered PMC wake interrupt
When using an edge triggered PMC wake interrupt for the mem_req->low
event, interrupts are missed when the GIC is suspended in idle. The
LIC does not support edge triggered interrupts so an alternate
implementation is used that utilizes level triggered interrupts
and toggles the active level. This approach reduces the number of
cases where the BBC is hibernating and the BBC frequency floor is
left enabled unnecessarily.

Bug 1329641

Change-Id: Ide55b0f2a9d42ed7c5b6c897b0f164ddcaa5da9f
Signed-off-by: Neil Patel <neilp@nvidia.com>
Reviewed-on: http://git-master/r/357842
Reviewed-by: Peter Hsiang <phsiang@nvidia.com>
Reviewed-by: Vinayak Pane <vpane@nvidia.com>
Reviewed-by: Prashant Malani <pmalani@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: Greg Heinrich <gheinrich@nvidia.com>
Reviewed-by: Steve Lin <stlin@nvidia.com>
Reviewed-by: Rick Song <ricks@nvidia.com>